+++ lib/views/git-tab-header-view.js
@@ -3,61 +3,18 @@ import PropTypes from 'prop-types';
 import path from 'path';
 import {CompositeDisposable} from 'atom';
 import {nullAuthor} from '../models/author';
+import {AuthorPropType} from '../prop-types';
 
 export default class GitTabHeaderView extends React.Component {
   static propTypes = {
-    currentWorkDir: PropTypes.string,
+    committer: AuthorPropType.isRequired,
 
+    // Workspace
+    workdir: PropTypes.string,
+    workdirs: PropTypes.arrayOf(PropTypes.string).isRequired,
+
+    // Event Handlers
     handleWorkDirSelect: PropTypes.func,
-    onDidChangeWorkDirs: PropTypes.func,
-    onDidUpdateRepo: PropTypes.func,
-    isRepoDestroyed: PropTypes.func.isRequired,
-    getCurrentWorkDirs: PropTypes.func.isRequired,
-    getCommitter: PropTypes.func.isRequired,
-  }
-
-  constructor(props) {
-    super(props);
-    this._isMounted = false;
-    this.state = {currentWorkDirs: [], committer: nullAuthor};
-    this.disposable = new CompositeDisposable();
-  }
-
-  static getDerivedStateFromProps(props, state) {
-    return {
-      ...state,
-      currentWorkDirs: props.getCurrentWorkDirs(),
-    };
-  }
-
-  componentDidMount() {
-    this._isMounted = true;
-    if (this.props.onDidChangeWorkDirs) {
-      this.disposable.add(this.props.onDidChangeWorkDirs(this.updateWorkDirs));
-    }
-    if (this.props.onDidUpdateRepo) {
-      this.disposable.add(this.props.onDidUpdateRepo(this.updateCommitter));
-    }
-    this.updateCommitter();
-  }
-
-  componentDidUpdate(prevProps) {
-    if (
-      prevProps.onDidChangeWorkDirs !== this.props.onDidChangeWorkDirs
-      || prevProps.onDidUpdateRepo !== this.props.onDidUpdateRepo
-    ) {
-      this.disposable.dispose();
-      this.disposable = new CompositeDisposable();
-      if (this.props.onDidChangeWorkDirs) {
-        this.disposable.add(this.props.onDidChangeWorkDirs(this.updateWorkDirs));
-      }
-      if (this.props.onDidUpdateRepo && !this.props.isRepoDestroyed()) {
-        this.disposable.add(this.props.onDidUpdateRepo(this.updateCommitter));
-      }
-    }
-    if (prevProps.getCommitter !== this.props.getCommitter) {
-      this.updateCommitter();
-    }
   }
 
   render() {
@@ -65,7 +22,7 @@ export default class GitTabHeaderView extends React.Component {
       <header className="github-Project">
         {this.renderCommitter()}
         <select className="github-Project-path input-select"
-          value={this.props.currentWorkDir ? path.normalize(this.props.currentWorkDir) : undefined}
+          value={this.props.workdir ? path.normalize(this.props.workdir) : undefined}
           onChange={this.props.handleWorkDirSelect ? this.props.handleWorkDirSelect : () => {}}>
           {this.renderWorkDirs()}
         </select>
@@ -75,22 +32,16 @@ export default class GitTabHeaderView extends React.Component {
 
   renderWorkDirs() {
     const workdirs = [];
-    for (const workdir of this.state.currentWorkDirs) {
+    for (const workdir of this.props.workdirs) {
       workdirs.push(<option key={workdir} value={path.normalize(workdir)}>{path.basename(workdir)}</option>);
     }
     return workdirs;
   }
 
-  updateWorkDirs = () => {
-    this.setState((state, props) => ({
-      currentWorkDirs: props.getCurrentWorkDirs(),
-    }));
-  }
-
   renderCommitter() {
-    const email = this.state.committer.getEmail();
-    const avatarUrl = this.state.committer.getAvatarUrl();
-    const name = this.state.committer.getFullName();
+    const email = this.props.committer.getEmail();
+    const avatarUrl = this.props.committer.getAvatarUrl();
+    const name = this.props.committer.getFullName();
 
     return (
       <img className="github-Project-avatar"
@@ -99,19 +50,5 @@ export default class GitTabHeaderView extends React.Component {
         alt={`${name}'s avatar`}
       />
     );
-  }
-
-  updateCommitter = async () => {
-    const committer = await this.props.getCommitter() || nullAuthor;
-    if (this._isMounted) {
-      this.setState({committer});
-    }
-  }
-
-  componentWillUnmount() {
-    this._isMounted = false;
-    if (this.disposable && this.disposable.dispose) {
-      this.disposable.dispose();
-    }
   }
 }
