+++ lib/views/git-tab-view.js
@@ -4,7 +4,7 @@ import cx from 'classnames';
 import {CompositeDisposable} from 'atom';
 
 import StagingView from './staging-view';
-import GitTabHeaderView from './git-tab-header-view';
+import GitTabHeaderController from '../controllers/git-tab-header-controller';
 import CommitController from '../controllers/commit-controller';
 import RecentCommitsController from '../controllers/recent-commits-controller';
 import RefHolder from '../models/ref-holder';
@@ -120,14 +120,18 @@ export default class GitTabView extends React.Component {
   renderHeader() {
     const {repository} = this.props;
     return (
-      <GitTabHeaderView
-        handleWorkDirSelect={e => this.props.changeWorkingDirectory(e.target.value)}
+      <GitTabHeaderController
+        getCommitter={repository.getCommitter.bind(repository)}
+        isRepoDestroyed={repository.isDestroyed.bind(repository)}
+        
+        // Workspace
         currentWorkDir={this.props.workingDirectoryPath}
         getCurrentWorkDirs={this.props.getCurrentWorkDirs}
+
+        // Event Handlers
         onDidChangeWorkDirs={this.props.onDidChangeWorkDirs}
-        getCommitter={repository.getCommitter.bind(repository)}
         onDidUpdateRepo={repository.onDidUpdate.bind(repository)}
-        isRepoDestroyed={repository.isDestroyed.bind(repository)}
+        handleWorkDirSelect={e => this.props.changeWorkingDirectory(e.target.value)}
       />
     );
   }
